Create parsermap_v2
This code will create a directory named "Reports" in the same location as your script. For each Serv.name value, it will generate a separate text file containing the corresponding results. The script appends results to existing files if they already exist, so if you run the script multiple times, the results will accumulate in the files.

Make sure you have the necessary permissions to create directories and files in the specified location. Also, ensure that the required gems (rubygems and nmap-parser) are installed in your Ruby environment.